aux: refactor visualisation.py into multiple modules
