+++ src/visualisation/geoviewer.py
@@ -1,21 +1,17 @@
-"""
-This module contains visualisation functions for GeoGraphs.
-"""
+"""This module contains the GeoGraphViewer to visualise GeoGraphs"""
+
 from __future__ import annotations
-from typing import Optional, List, Tuple, Callable, Dict, Union
+from typing import List, Dict, Union
 
 import folium
 import pandas as pd
-import geopandas as gpd
-import networkx as nx
-import shapely.geometry
 import ipywidgets as widgets
 import ipyleaflet
 import traitlets
 
-
-from src.constants import CHERNOBYL_COORDS_WGS84, WGS84, UTM35N, CEZ_DATA_PATH
+from src.constants import CHERNOBYL_COORDS_WGS84, WGS84
 from src.models import geograph, metrics
+from src.visualisation import folium_utils, graph_utils
 
 
 class GeoGraphViewer(ipyleaflet.Map):
@@ -125,7 +121,7 @@ class GeoGraphViewer(ipyleaflet.Map):
 
             is_habitat = idx > 0
 
-            nodes, edges = create_node_edge_geometries(nx_graph)
+            nodes, edges = graph_utils.create_node_edge_geometries(nx_graph)
             graph_geometries = pd.concat([edges, nodes]).reset_index()
             graph_geo_data = ipyleaflet.GeoData(
                 geo_dataframe=graph_geometries.to_crs(WGS84),
@@ -547,197 +543,6 @@ class GeoGraphViewer(ipyleaflet.Map):
         self.add_control(ipyleaflet.ScaleControl(position="bottomleft"))
 
 
-class VisibilityWidget(widgets.Box):
-    """Widget to control visibility of graphs in GeoGraphViewer."""
-
-    log_out = widgets.Output(layout={"border": "1px solid black"})
-
-    @log_out.capture()
-    def __init__(self, viewer: GeoGraphViewer) -> None:
-        self.viewer = viewer
-        self.viewer.hidde_all_layers()
-
-        # Setting the order and current view of graph and map
-        # with the latter two as traits
-        self.add_traits(
-            graph_index=traitlets.Int().tag(sync=True),
-            map_index=traitlets.Int().tag(sync=True),
-            current_graph=traitlets.Unicode().tag(sync=True),
-            current_map=traitlets.Unicode().tag(sync=True),
-        )
-
-        self.graph_index = 0
-        self.map_index = 0
-        self.graph_names = list(viewer.layer_dict["graphs"].keys())
-        self.map_names = list(viewer.layer_dict["maps"].keys())
-
-        # If the indices change, the current_map and current_graph
-        # change automatically as well.
-        self._observe_index(change=None)
-        self.observe(self._observe_index, names=["graph_index", "map_index"])
-
-        # creating widget
-        widget = self.assemble_widgets()
-
-        super().__init__([widget])
-
-    @log_out.capture()
-    def assemble_widgets(self) -> widgets.Widget:
-        """Assemble all sub-widgets making up VisibilityWidget into layout.
-
-        Returns:
-            widgets.Widget: final widget to be added to GeoGraphViewer
-        """
-
-        graph_selection = self.create_graph_selection()
-        view_buttons = self.create_visibility_buttons()
-
-        widget = widgets.VBox([graph_selection, view_buttons])
-
-        return widget
-
-    @log_out.capture()
-    def _observe_index(self, change):
-        """Set current graph and map based on indices"""
-        print("Changed index:", change)
-        self.current_graph = self.graph_names[self.graph_index]
-        self.current_map = self.map_names[self.map_index]
-
-    @log_out.capture()
-    def create_graph_selection(self) -> widgets.RadioButtons:
-        """Create radio buttons to enable graph selection.
-
-        Returns:
-            widgets.RadioButtons: buttons to select graph
-        """
-
-        graph_list = []
-        for idx, graph_name in enumerate(self.graph_names):
-            graph_str = graph_name
-            graph_list.append((graph_str, idx))
-
-        radio_buttons = widgets.RadioButtons(options=graph_list, description="")
-        widgets.link((radio_buttons, "value"), (self, "graph_index"))
-
-        return radio_buttons
-
-    @log_out.capture()
-    def create_visibility_buttons(self) -> widgets.Box:
-        """Create buttons that toggle the visibility of current graph and map.
-
-        The visibility of the current graph (set in self.current_graph), its subparts
-        and the map (set in self.current_map) can be manipulated with the returned
-        buttons. Separate buttons for the polygons and the components of the graph are
-        included in the returned box.
-
-        Returns:
-            widgets.Box: box with button widgets
-        """
-
-        # Creating view buttons
-        btn_layout = widgets.Layout(width="115px")
-        view_graph_btn = widgets.ToggleButton(
-            description="Graph",
-            tooltip="View graph",
-            icon="project-diagram",
-            layout=btn_layout,
-        )
-        view_pgon_btn = widgets.ToggleButton(
-            description="Polygons",
-            disabled=False,
-            button_style="",  # 'success', 'info', 'warning', 'danger' or ''
-            tooltip="View graph",
-            icon="shapes",
-            layout=btn_layout,
-        )
-        view_components_btn = widgets.ToggleButton(
-            description="Components",
-            button_style="",  # 'success', 'info', 'warning', 'danger' or ''
-            tooltip="View graph",
-            icon="circle",
-            layout=btn_layout,
-        )
-        view_map_btn = widgets.ToggleButton(
-            description="Map",
-            button_style="",  # 'success', 'info', 'warning', 'danger' or ''
-            tooltip="View graph",
-            icon="globe-africa",
-            layout=btn_layout,
-        )
-
-        button_list = [
-            view_graph_btn,
-            view_pgon_btn,
-            view_components_btn,
-            view_map_btn,
-        ]
-
-        # Adding additional traits for callback
-        for button, layer_subtype in zip(
-            button_list, ["graph", "pgons", "components", "map"]
-        ):
-            button.add_traits(
-                layer_type=traitlets.Unicode().tag(sync=True),
-                layer_subtype=traitlets.Unicode().tag(sync=True),
-                layer_name=traitlets.Unicode().tag(sync=True),
-            )
-            if layer_subtype == "map":
-                button.layer_type = "maps"
-                button.layer_name = self.current_map
-                # If current map changes the function of this button changes
-                widgets.dlink((self, "current_map"), (button, "layer_name"))
-            else:
-                button.layer_type = "graphs"
-                button.layer_name = self.current_graph
-                widgets.dlink((self, "current_graph"), (button, "layer_name"))
-
-            button.layer_subtype = layer_subtype
-
-            button.observe(self._handle_view, names=["value", "layer_name"])
-
-        view_graph_btn.value = True
-        view_pgon_btn.value = True
-        view_map_btn.value = True
-
-        buttons = widgets.VBox(
-            [widgets.HBox(button_list[:2]), widgets.HBox(button_list[2:])]
-        )
-
-        return buttons
-
-    @log_out.capture()
-    def _handle_view(self, change: Dict) -> None:
-        """Callback function for trait events in view buttons"""
-        owner = change.owner  # Button that is clicked or changed
-
-        print("Detected button change:", change)
-
-        # Button clicked
-        if change.name == "value":
-            active = change.new
-            self.viewer.set_layer_visibility(
-                owner.layer_type, owner.layer_name, owner.layer_subtype, active
-            )
-            self.viewer.layer_update()
-
-        # Layer that the button is assigned is changed
-        elif change.name == "layer_name":
-            new_layer_name = change.new
-            old_layer_name = change.old
-
-            # remove old layer
-            self.viewer.set_layer_visibility(
-                owner.layer_type, old_layer_name, owner.layer_subtype, False
-            )
-            # view new layer
-            self.viewer.set_layer_visibility(
-                owner.layer_type, new_layer_name, owner.layer_subtype, owner.value
-            )
-            # Note: there is a potential for speed improvement by not updating map
-            # layers for each button separately, as is done here.
-            self.viewer.layer_update()
-
-
 class FoliumGeoGraphViewer:
     """Class for viewing GeoGraph object without ipywidgets"""
 
@@ -775,229 +580,6 @@ class FoliumGeoGraphViewer:
         Args:
             graph (geograph.GeoGraph): GeoGraph to be added.
         """
-        self.widget = add_graph_to_folium_map(folium_map=self.widget, graph=graph.graph)
-
-
-def add_graph_to_folium_map(
-    folium_map: folium.Map = None,
-    polygon_gdf: gpd.GeoDataFrame = None,
-    color_column: str = "index",
-    graph: Optional[nx.Graph] = None,
-    name: str = "data",
-    folium_tile_list: Optional[List[str]] = None,
-    location: Tuple[float, float] = CHERNOBYL_COORDS_WGS84,
-    crs: str = UTM35N,
-    add_layer_control: bool = False,
-) -> folium.Map:
-    """Create a visualisation map of the given polygons and `graph` in folium.
-
-    The polygons in `polygon_gdf` and `graph` are displayed on a folum map.
-    It is intended that the graph was build from `polygon_gdf`, but it is not required.
-    If given `map`, it will be put on this existing folium map.
-
-    Args:
-        folium_map (folium.Map, optional): map to add polygons and graph to.
-            Defaults to None.
-        polygon_gdf (gpd.GeoDataFrame, optional): data containing polygon.
-            Defaults to None.
-        color_column (str, optional): column in polygon_gdf that determines which color
-            is given to each polygon. Can be categorical values. Defaults to "index".
-        graph (Optional[nx.Graph], optional): graph to be plotted. Defaults to None.
-        name (str, optional): prefix to all the folium layer names shown in layer
-            control of map (if added). Defaults to "data".
-        folium_tile_list (Optional[List[str]], optional): list of folium.Map tiles to be
-            add to the map. See folium.Map docs for options. Defaults to None.
-        location (Tuple[float, float], optional): starting location in WGS84 coordinates
-            Defaults to CHERNOBYL_COORDS_WGS84.
-        crs (str, optional): coordinates reference system to be used.
-            Defaults to UTM35N.
-        add_layer_control (bool, optional): whether to add layer controls to map.
-            Warning: only use this when you don't intend to add any additional data
-            after calling this function to the map. May cause bugs otherwise.
-            Defaults to False.
-
-    Returns:
-        folium.Map: map with polygons and graph displayed as described
-    """
-
-    if folium_tile_list is None:
-        folium_tile_list = ["OpenStreetMap"]
-
-    if folium_map is None:
-        folium_map = folium.Map(location, zoom_start=8, tiles=folium_tile_list.pop(0))
-
-    # Adding standard folium raster tiles
-    for tiles in folium_tile_list:
-        # special esri satellite data case
-        if tiles == "esri":
-            folium.TileLayer(
-                tiles=(
-                    "https://server.arcgisonline.com/ArcGIS/rest/"
-                    "services/World_Imagery/MapServer/tile/{z}/{y}/{x}"
-                ),
-                attr="esri",
-                name="esri satellite",
-                overlay=False,
-                control=True,
-            ).add_to(folium_map)
-        else:
-            folium.TileLayer(tiles=tiles).add_to(folium_map)
-
-    # Adding polygon data
-    if polygon_gdf is not None:
-        # creating a color index that maps each category
-        # in the color_column to an integer
-        polygon_gdf["index"] = polygon_gdf.index
-        polygon_gdf["color_index"] = (
-            polygon_gdf[color_column].astype("category").cat.codes.astype("int64")
+        self.widget = folium_utils.add_graph_to_folium_map(
+            folium_map=self.widget, graph=graph.graph
         )
-
-        choropleth = folium.Choropleth(
-            polygon_gdf,
-            data=polygon_gdf,
-            key_on="feature.properties.index",
-            columns=["index", "color_index"],
-            fill_color="YlOrBr",
-            name=name + "_polygons",
-        )
-        choropleth = remove_choropleth_color_legend(choropleth)
-        choropleth.add_to(folium_map)
-
-        # adding popup markers with class name
-        folium.features.GeoJsonPopup(fields=[color_column], labels=True).add_to(
-            choropleth.geojson
-        )
-
-    # Adding graph data
-    if graph is not None:
-        node_gdf, edge_gdf = create_node_edge_geometries(graph, crs=crs)
-
-        # add graph edges to map
-        if not edge_gdf.empty:
-            edges = folium.features.GeoJson(
-                edge_gdf,
-                name=name + "_graph_edges",
-                style_function=get_style_function("#dd0000"),
-            )
-            edges.add_to(folium_map)
-
-        # add graph nodes/vertices to map
-        node_marker = folium.vector_layers.Circle(radius=100, color="black")
-        nodes = folium.features.GeoJson(
-            node_gdf, marker=node_marker, name=name + "_graph_vetrices"
-        )
-        nodes.add_to(folium_map)
-
-    if add_layer_control:
-        folium.LayerControl().add_to(folium_map)
-
-    return folium_map
-
-
-def create_node_edge_geometries(
-    graph: nx.Graph, crs: str = UTM35N
-) -> Tuple[gpd.GeoDataFrame, gpd.GeoDataFrame]:
-    """Create node and edge geometries for the networkx graph G.
-
-    Returns node and edge geometries in two GeoDataFrames. The output can be used for
-    plotting a graph.
-
-    Args:
-        graph (nx.Graph): graph with nodes and edges
-        crs (str, optional): coordinate reference system. Defaults to UTM35N.
-
-    Returns:
-        Tuple[gpd.GeoDataFrame, gpd.GeoDataFrame]: dataframes of nodes and edges
-            respectively.
-    """
-
-    node_gdf = gpd.GeoDataFrame(columns=["id", "geometry"])
-    rep_points = graph.nodes(data="rep_point")
-    for idx, rep_point in rep_points:
-        node_gdf.loc[idx] = [idx, rep_point]
-
-    edge_gdf = gpd.GeoDataFrame(columns=["id", "geometry"])
-    for idx, (node_a, node_b) in enumerate(graph.edges()):
-        point_a = rep_points[node_a]
-        point_b = rep_points[node_b]
-        line = shapely.geometry.LineString([point_a, point_b])
-
-        edge_gdf.loc[idx] = [idx, line]
-
-    node_gdf = node_gdf.set_crs(crs)
-    edge_gdf = edge_gdf.set_crs(crs)
-
-    return node_gdf, edge_gdf
-
-
-def get_style_function(color: str = "#ff0000") -> Callable[[], dict]:
-    """Return lambda function that returns a dict with the `color` given.
-
-    The returned lambda function can be used as a style function for folium.
-
-    Args:
-        color (str, optional): color to be used in dict. Defaults to "#ff0000".
-
-    Returns:
-        Callable[[], dict]: style function
-    """
-
-    return lambda x: {"fillColor": color, "color": color}
-
-
-def add_cez_to_map(
-    folium_map: folium.Map,
-    exclusion_json_path: Optional[str] = CEZ_DATA_PATH,
-    add_layer_control: bool = False,
-) -> folium.Map:
-    """Add polygons of the Chernobyl Exclusion Zone (CEZ) to a folium map.
-
-    Args:
-        folium_map (folium.Map): [description]
-        exclusion_json_path (Optional[str], optional): path to the json file containing
-            the CEZ polygons. Defaults to CEZ_DATA_PATH which requires access to
-            the Jasmin servers and relevant shared workspaces.
-        add_layer_control (bool, optional): whether to add layer controls to map.
-            Warning: only use this when you don't intend to add any additional data
-            after calling this function to the map. May cause bugs otherwise.
-            Defaults to False.
-    Returns:
-        folium.Map: map with CEZ polygons added
-    """
-
-    exc_data = gpd.read_file(exclusion_json_path)
-
-    colors = ["#808080", "#ffff99", "#ff9933", "#990000", "#ff0000", "#000000"]
-
-    for index, row in exc_data.iterrows():
-        folium.GeoJson(
-            row["geometry"],
-            name=row["name"],
-            style_function=get_style_function(colors[index]),
-        ).add_to(folium_map)
-
-    if add_layer_control:
-        folium.LayerControl().add_to(folium_map)
-
-    return folium_map
-
-
-def remove_choropleth_color_legend(
-    choropleth_map: folium.features.Choropleth,
-) -> folium.features.Choropleth:
-    """Remove color legend from Choropleth folium map.
-
-    Solution proposed by `nhpackard` in the following GitHub issue in the folium repo:
-    https://github.com/python-visualization/folium/issues/956
-
-    Args:
-        choropleth_map (folium.features.Choropleth): a Choropleth map
-
-    Returns:
-        folium.features.Choropleth: the same map without color legend
-    """
-    for key in choropleth_map._children:  # pylint: disable=protected-access
-        if key.startswith("color_map"):
-            del choropleth_map._children[key]  # pylint: disable=protected-access
-
-    return choropleth_map
